+++ docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.1.0
+### Added
+- JSON serialization for `CosmicTime`
+
+### Changed
+- Changed JSON serialization from string-based to property-based
+
 ## 1.0.0
 ### Added
 - Initial release